Millionaire

Description

Documentation

type DP = Map (NumTry, Level) Prob

type NumTry = Int

type Level = Int

type Prob = Double

type Money = Double

main :: IO ()

Takes the total number of rounds, the winning probability for each round (constant), and the initial amount of money, and outputs the probability of winning in the end.

The input must have the form:

M P X

solve

Arguments

:: Prob
The probability of winning each round

-> (Money, Money)
(Goal amount, Initial amount of money)

-> NumTry
The total number of rounds

-> Prob
The probability of winning the whole game

Calculates the probability of winning the game given the probability of winning each round, the goal amount, and the initial amount of money.

makeBet

Arguments

:: Prob
The probability of winning each round

-> DP
The dp table

-> (NumTry, Level)
(Remaining number of rounds, Current level)

-> DP
The updated dp table

Makes a bet based on the current value and updates the dp table

getBetResults

Arguments

:: NumTry
Remaining number of rounds

-> Prob
Current expected value

-> DP
dp table

-> (Prob, Level)
(Probability, Current level)

-> DP
The updated dp table

Updates the dp table with the result of a bet

getSchedule

Arguments

:: NumTry
The total number of rounds

-> [(NumTry, Level)]
The list of all combinations of number of rounds and levels

Calculates all combinations of number of rounds and levels from the total number of rounds

initDPTable

Arguments

:: (Money, Money)
(Goal amount, Initial amount of money)

-> NumTry
The total number of rounds

-> DP
The initialized dp table

Initializes the dp table

getLevel

Arguments

:: (Money, Money)
(Goal amount, Initial amount of money)

-> NumTry
The total number of rounds

-> Level
The level of the money

Calculates the level of the money

goal :: Money

The goal amount of money