POJ1182

Description

Documentation

type UF = DisjointSet (Index, Char)

type State = (Int, UF)

type Index = Int

type IdPair = (Index, Index)

main :: IO ()

The main function

The input must have form;

N K

type1 x1 y1

type2 x2 y2

...

typeK xK yK

initUF

Arguments

:: Int
The number of individuals

-> UF
The initial disjoint set

Initializes the disjoint set

dealQuery

Arguments

:: Int
The number of individuals

-> State
The current number of invalid queries and the disjoint set

-> [Int]
The query

-> State
The updated number of invalid queries and the disjoint set

deals with the query and updates the state

process

Arguments

:: [(Char, Char)]
Char pairs that define which chars to be connected

-> [(Char, Char)]
Char pairs that define which chars are not to be connected

-> State
The current state

-> IdPair
The indecies that are queried

-> State
The updated state

Processes the first type of query

update

Arguments

:: [(Char, Char)]
Char pairs that define which chars to be connected

-> IdPair
The indecies that are queried

-> UF
The current disjoint set

-> UF
The updated disjoint set

Updates the disjoint set

check

Arguments

:: [(Char, Char)]
Char pairs that define which chars are not to be connected

-> IdPair
The indecies that are queried

-> UF
The disjoint set

-> Bool
If the query is consistent to the history

Checks if the new query is consistent to the history

ref1 :: [(Char, Char)]

Chars that are to be connected for type1 queries

ref1_anti :: [(Char, Char)]

Chars that are not to be connected for type1 queries

ref2 :: [(Char, Char)]

Chars that are to be connected for type2 queries

ref2_anti :: [(Char, Char)]

Chars that are not to be connected for type2 queries